+++ src/main/java/edu/school21/cinema/repository/UserRepository.java
@@ -2,9 +2,7 @@ package edu.school21.cinema.repository;
 
 import edu.school21.cinema.model.User;
 
-import java.util.List;
-
 public interface UserRepository {
-    Long save(User user);
+    Integer save(User user);
     User getUserByEmail(String email);
 }
